requirements-dev.txt
pytest==5.4.1
pytest-vcr==1.0.2
pycodestyle==2.5.0
pytest-cov==2.8.1
black==19.10b0
isort==4.3.21
flake8==3.7.9
mypy==0.770
coveralls==2.0.0
pytest-chalice==0.0.4
